delete and update records

+++ ccm_app/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
-from .forms import SignUpForm
+from .forms import SignUpForm, AddRecordForm
 from .models import Record
 
 
@@ -59,10 +59,43 @@ def payment_record(request, pk):
 
 def delete_record(request, pk):
     if request.user.is_authenticated:
-        delete_it = Record.objects.get(id=pk)
-        delete_it.delete()
-        messages.success(request, 'Record has been deleted!')
-        return redirect('home')
+        if request.user.is_superuser:
+            delete_it = Record.objects.get(id=pk)
+            delete_it.delete()
+            messages.success(request, 'Record has been deleted!')
+            return redirect('home')
+        else:
+            messages.success(request, 'You must be an admin to delete records!')
+            return redirect('home')
     else:
         messages.success(request, 'You must be logged in to delete records!')
+        return redirect('home')
+
+
+def add_record(request):
+    form = AddRecordForm(request.POST or None)
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Record has been added!')
+                return redirect('home')
+        return render(request, 'add_record.html', {'form': form})
+    else:
+        messages.success(request, 'You must be logged in to add records!')
+        return redirect('home')
+
+
+def update_record(request, pk):
+    if request.user.is_authenticated:
+        current_record = Record.objects.get(id=pk)
+        form = AddRecordForm(request.POST or None, instance=current_record)
+        if request.method == 'POST':
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Record has been updated!')
+                return redirect('home')
+        return render(request, 'update_record.html', {'form': form})
+    else:
+        messages.success(request, 'You must be logged in to update records!')
         return redirect('home')
